mysite/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-decouple
django-cors-headers
environs